adjust class position
