+++ core/src/main/java/com/alibaba/fastjson2/filter/PropertyPreFilter.java
@@ -2,9 +2,6 @@ package com.alibaba.fastjson2.filter;
 
 import com.alibaba.fastjson2.JSONWriter;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public interface PropertyPreFilter
         extends Filter {
     boolean process(JSONWriter writer, Object source, String name);
@@ -16,34 +13,5 @@ public interface PropertyPreFilter
         }
 
         return new CompositePropertyPreFilter(before, after);
-    }
-
-    final class CompositePropertyPreFilter
-            implements PropertyPreFilter {
-        private final List<PropertyPreFilter> filters = new ArrayList<>();
-
-        public CompositePropertyPreFilter(PropertyPreFilter... filters) {
-            for (PropertyPreFilter filter : filters) {
-                if (filter != null) {
-                    this.filters.add(filter);
-                }
-            }
-        }
-
-        public void add(PropertyPreFilter filter) {
-            if (filter != null) {
-                this.filters.add(filter);
-            }
-        }
-
-        @Override
-        public boolean process(JSONWriter writer, Object source, String name) {
-            for (PropertyPreFilter filter : filters) {
-                if (!filter.process(writer, source, name)) {
-                    return false;
-                }
-            }
-            return true;
-        }
     }
 }
